Basisklasse für WireMockTests extrahiert
